# Remove n+1 queries when fetching association counts on /admin/enterprises