--- a/app/queries/batch_enterprise_enterprise_fees_count_query.rb
+++ b/app/queries/batch_enterprise_enterprise_fees_count_query.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class BatchEnterpriseEnterpriseFeesCountQuery
+  def self.call(enterprise_ids)
+    EnterpriseFee.where(enterprise_id: enterprise_ids).group(:enterprise_id).count
+  end
+end