--- a/app/queries/batch_enterprise_producer_properties_count_query.rb
+++ b/app/queries/batch_enterprise_producer_properties_count_query.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class BatchEnterpriseProducerPropertiesCountQuery
+  def self.call(enterprise_ids)
+    ProducerProperty.where(producer_id: enterprise_ids).group(:producer_id).unscope(:order).count
+  end
+end