When the `/admin/enterprises` page is loaded the following queries are called in `app/views/admin/enterprises/_actions.html.haml` for every enterprise.

```
enterprise.enterprise_fees.count
enterprise.producer_properties.count
enterprise.payment_methods.count
enterprise.shipping_methods.count
```

This is to improve performance for #8027. Instead of running a query for each enterprise individually, for example to count the shipping methods that enterprise has, it runs a single query to count the shipping methods for all the enterprises loaded on the page all at once.

Before

```
Completed 200 OK in 3670ms (Views: 2688.4ms | ActiveRecord: 959.7ms | Allocations: 1664423)
Completed 200 OK in 3322ms (Views: 2402.5ms | ActiveRecord: 887.2ms | Allocations: 1671457)
Completed 200 OK in 3404ms (Views: 2513.2ms | ActiveRecord: 868.6ms | Allocations: 1663917)
```

After

```
Completed 200 OK in 1832ms (Views: 1397.0ms | ActiveRecord: 402.6ms | Allocations: 1014829)
Completed 200 OK in 2121ms (Views: 1600.5ms | ActiveRecord: 446.0ms | Allocations: 1023386)
Completed 200 OK in 1851ms (Views: 1408.8ms | ActiveRecord: 408.0ms | Allocations: 1015167)
```
#### What should we test?

Test that /admin/enterprises page loads faster than before.

#### Release notes

Remove n+1 queries when fetching association counts on /admin/enterprises.

Changelog Category: User facing changes