@cillian 2021-08-31
Thanks @Matt-Yorkley I didn't think of pagination, that sounds a lot simpler anyway. I close this one.